chore: extract common code
